modification des controller concernant les delete

+++ src/Controller/AdminActivityController.php
@@ -9,11 +9,14 @@ use \FilesystemIterator;
 class AdminActivityController extends AbstractController
 {
 
-    public function deleteActivity(int $id): void
+    public function deleteActivity(): void
     {
-        $activityManager = new ActivityManager();
-        $activityManager->delete($id);
-        header('Location:/admin/index');
+        if (!empty($_POST['id'])) {
+            $id = $_POST['id'];
+            $activityManager = new ActivityManager();
+            $activityManager->delete($id);
+            header('Location:/admin/index');
+        }
     }
 
     public function addActivity(string $message = "")
